+++ src/Endpoint/ContactImportsEndpoint.php
@@ -9,6 +9,7 @@ use DigitalCz\DigiSign\Endpoint\Traits\DeleteEndpointTrait;
 use DigitalCz\DigiSign\Endpoint\Traits\GetEndpointTrait;
 use DigitalCz\DigiSign\Endpoint\Traits\UpdateEndpointTrait;
 use DigitalCz\DigiSign\Resource\ContactImport;
+use DigitalCz\DigiSign\Resource\ContactImportProgress;
 
 /**
  * @extends ResourceEndpoint<ContactImport>
@@ -32,5 +33,21 @@ final class ContactImportsEndpoint extends ResourceEndpoint
     public function items(ContactImport|string $id): ContactImportItemsEndpoint
     {
         return new ContactImportItemsEndpoint($this, $id);
+    }
+
+    public function progress(ContactImport|string $id): ContactImportProgress
+    {
+        return $this->createResource(
+            $this->getRequest('/{id}/progress', ['id' => $id]),
+            ContactImportProgress::class,
+        );
+    }
+
+    public function start(ContactImport|string $id): ContactImportProgress
+    {
+        return $this->createResource(
+            $this->postRequest('/{id}/start', ['id' => $id]),
+            ContactImportProgress::class,
+        );
     }
 }